+++ src/main/kotlin/com/sakurasedaia/blenderextensions/run/BlenderRunConfigurationType.kt
@@ -14,15 +14,47 @@ class BlenderRunConfigurationType : ConfigurationType {
     override fun getIcon(): Icon = BlenderIcons.Blender
     override fun getId(): String = "BLENDER_RUN_CONFIGURATION"
     override fun getConfigurationFactories(): Array<ConfigurationFactory> = arrayOf(
-        BlenderConfigurationFactory(this),
+        BlenderTestingConfigurationFactory(this),
+        BlenderBuildConfigurationFactory(this),
+        BlenderValidateConfigurationFactory(this),
         BlenderCommandConfigurationFactory(this)
     )
 }
 
+class BlenderTestingConfigurationFactory(type: ConfigurationType) : ConfigurationFactory(type) {
+    override fun createTemplateConfiguration(project: Project): RunConfiguration =
+        BlenderRunConfiguration(project, this, "Testing")
+
+    override fun getId(): String = "BlenderTestingConfigurationFactory"
+    override fun getOptionsClass(): Class<out BaseState> = BlenderRunConfigurationOptions::class.java
+}
+
+class BlenderBuildConfigurationFactory(type: ConfigurationType) : ConfigurationFactory(type) {
+    override fun createTemplateConfiguration(project: Project): RunConfiguration {
+        val config = BlenderRunConfiguration(project, this, "Build")
+        config.getOptions().blenderCommand = "extensions build"
+        return config
+    }
+
+    override fun getId(): String = "BlenderBuildConfigurationFactory"
+    override fun getOptionsClass(): Class<out BaseState> = BlenderRunConfigurationOptions::class.java
+}
+
+class BlenderValidateConfigurationFactory(type: ConfigurationType) : ConfigurationFactory(type) {
+    override fun createTemplateConfiguration(project: Project): RunConfiguration {
+        val config = BlenderRunConfiguration(project, this, "Validate")
+        config.getOptions().blenderCommand = "extensions validate"
+        return config
+    }
+
+    override fun getId(): String = "BlenderValidateConfigurationFactory"
+    override fun getOptionsClass(): Class<out BaseState> = BlenderRunConfigurationOptions::class.java
+}
+
 class BlenderCommandConfigurationFactory(type: ConfigurationType) : ConfigurationFactory(type) {
     override fun createTemplateConfiguration(project: Project): RunConfiguration {
-        val config = BlenderRunConfiguration(project, this, "Blender Command")
-        config.getOptions().blenderCommand = "extension validate"
+        val config = BlenderRunConfiguration(project, this, "Command")
+        config.getOptions().blenderCommand = ""
         return config
     }